Removed improviment #51
Removed the improvement of information "it is cheaper to buy more, in the next price break" because of the compatibility of Xlxswrite if excel.
Microsoft Excel do not recognize arrays `{}` in the formula of conditional format, just LibreOffice does.
This feature was removed until now to keep compatibility between systems. Waiting better idea in in the #51 discussion.