Macro to the KiCost
This macro in bash (Linux prompt language, but also available in Windows 10 via the Microsoft update) test all the XML BOM files present in the folder.
If finish the script, all the passed all print this in the prompt. If don't, return error.
I am using to validate my last changes in the KiCost.