Issue #84 and other improvements
Now the spreadsheet order the groups by the defined BOM_ORDER definition;
`altium` was moved to `eda_tools` with `__init__.py` having the BOM_ORDER definition (to be used in all eda XML read routines). Is intention is in the future create a `kicad.py`module and remove the functions from `kicost.py`;
Multiparts is working for unitary values (recognize the values, but for now, not multiply in the spreadsheet).