Include variant-specific user-requested fields.
When old-style variants (i.e. fields with names like kicost.V1:manf)
are used to assign different part numbers based on the variant selection,
and variant-specific values are assigned to fields requested for inclusion
with the --fields option (e.g., assigning different internal part numbers
with fields like kicost.V1:mypartnum), the correct user-requested
field values were not consistently included in BOM lines. This change
makes sure that user-requested fields are not renamed with a "local:"
prefix, so that the normal variant-specific selection of field
values will be applied to those user-requested fields.