Replaced CurrencyConverter
This patch adds a small internal module to remove the CurrencyConverter
dependency.
The replacement always does a request for the exchange rates. It
doesn't rely on a hardcoded database that gets outdated as soon
as we do a release.
It has a fallback in case the request fail.
The patch also adds a tool to update the fallback before a release.