Update eda_kicad.py to check for the property "dnp" for components
In the function extract_fields, added a loop to check for the property "dnp" in components.

In KiCad 7.0, eeshema has an attribute "Do not populate" when editing a component properties. When enabled, the component will appear with an X overlaid on the component.  However, this gets added as a property in the .xml generated by kicad, not a field.

Tested on an in-house schematic.